requirements.txt
chainer==1.24.0
filelock==3.0.10
nose==1.3.7
numpy==1.13.3
opencv-python==3.2.0.8
Pillow==5.1.0
protobuf==3.6.1
PyQt5==5.10.1
QtPy==1.6.0
sip==4.19.8
six==1.12.0
